moved the one function in load_ptyr.py to ptycho_utils.py
